requirements.txt
arrow==1.2.2
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.3
django-geojson==3.2.1
django-import-export==2.7.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
jsonfield==3.1.0
MarkupPy==1.14
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
